replace DashboardLinks with refactored DashLinks
